Restructure mirage module
