+++ src/mirage/index.ts
@@ -1,3 +1,8 @@
-export * from './middleware';
-export * from './resolvers';
-export { MirageGraphQLMapper } from './mapper';
+export { MirageGraphQLMapper } from './mapper/mapper';
+export { patchAutoFieldResolvers } from './middleware/patch-auto-field-resolvers';
+export { patchAutoTypeResolvers } from './middleware/patch-auto-type-resolvers';
+export { patchAutoResolvers } from './middleware/patch-auto-resolvers';
+export { mirageRootQueryResolver } from './resolver/root-query';
+export { mirageObjectResolver } from './resolver/object';
+export { mirageUnionResolver } from './resolver/union';
+export { mirageInterfaceResolver } from './resolver/interface';
