--- src/mirage/middleware/index.ts
@@ -1,3 +1,0 @@
-export { patchAutoFieldResolvers as patchModelTypes } from './patch-auto-field-resolvers';
-export { patchAutoTypeResolvers as patchUnionsInterfaces } from './patch-auto-type-resolvers';
-export { patchAutoResolvers } from './patch-auto-resolvers';
