+++ src/mirage/resolver/object.ts
@@ -1,11 +1,12 @@
 import { Resolver, ResolverInfo, ResolverParent } from '../../types';
-import { MirageGraphQLMapper } from '../mapper';
+import { MirageGraphQLMapper } from '../mapper/mapper';
 import { relayPaginateNodes } from '../../relay/utils';
-import { mirageCursorForNode, ObjectResolverMatch, AutoResolverErrorMeta } from './helpers';
+import { mirageCursorForNode } from './utils';
 import { extractDependencies } from '../../resolver/extract-dependencies';
-import { AutoResolverError } from './auto-resolver-error';
+import { AutoResolverError } from '../auto-resolver-error';
 import { unwrap } from '../../graphql/utils';
 import { coerceToList, coerceReturnType } from '../../resolver/utils';
+import { ObjectResolverMatch, AutoResolverErrorMeta } from '../types';
 
 function findMatchingFieldForObjectParent({
   mirageMapper,
