+++ src/mirage/resolver/root-query.ts
@@ -1,18 +1,14 @@
 import { Server as MirageServer } from 'miragejs';
 import { isScalarType } from 'graphql';
 import { Resolver } from '../../types';
-import { MirageGraphQLMapper } from '../mapper';
+import { MirageGraphQLMapper } from '../mapper/mapper';
 import { relayPaginateNodes } from '../../relay/utils';
-import {
-  cleanRelayConnectionName,
-  mirageCursorForNode,
-  RootQueryResolverMatch,
-  AutoResolverErrorMeta,
-} from './helpers';
+import { cleanRelayConnectionName, mirageCursorForNode } from './utils';
 import { extractDependencies } from '../../resolver/extract-dependencies';
-import { AutoResolverError } from './auto-resolver-error';
+import { AutoResolverError } from '../auto-resolver-error';
 import { unwrap } from '../../graphql/utils';
 import { coerceToList, coerceReturnType } from '../../resolver/utils';
+import { RootQueryResolverMatch, AutoResolverErrorMeta } from '../types';
 
 function findMatchingModelsForType({
   type,
