--- src/mirage/resolvers/index.ts
@@ -1,3 +1,0 @@
-export { mirageInterfaceResolver } from './interface';
-export { mirageObjectResolver } from './object';
-export { mirageUnionResolver } from './union';
