+++ test/unit/mirage/resolvers/relay.test.ts
@@ -1,4 +1,4 @@
-import { mirageObjectResolver } from '../../../../src/mirage/resolvers/object';
+import { mirageObjectResolver } from '../../../../src/mirage/resolver/object';
 import { expect } from 'chai';
 import { buildSchema, GraphQLObjectType, GraphQLResolveInfo } from 'graphql';
 import { Model, Server, hasMany, ModelInstance, Registry } from 'miragejs';
